requirements.txt
ebaysdk
python-dotenv